Move block types to executor
