+++ core/lib/vm_executor/src/oneshot/block.rs
@@ -1,0 +1,353 @@
+use anyhow::Context;
+use zksync_dal::{Connection, Core, CoreDal, DalError};
+use zksync_multivm::{
+    interface::{L1BatchEnv, L2BlockEnv, OneshotEnv, StoredL2BlockEnv, SystemEnv, TxExecutionMode},
+    vm_latest::constants::BATCH_COMPUTATIONAL_GAS_LIMIT,
+};
+use zksync_types::{
+    api,
+    block::{unpack_block_info, L2BlockHasher},
+    fee_model::BatchFeeInput,
+    AccountTreeId, L1BatchNumber, L2BlockNumber, L2ChainId, ProtocolVersionId, StorageKey, H256,
+    SYSTEM_CONTEXT_ADDRESS, SYSTEM_CONTEXT_CURRENT_L2_BLOCK_INFO_POSITION,
+    SYSTEM_CONTEXT_CURRENT_TX_ROLLING_HASH_POSITION, ZKPORTER_IS_AVAILABLE,
+};
+use zksync_utils::{h256_to_u256, time::seconds_since_epoch};
+
+use super::MultiVMBaseSystemContracts;
+
+/// Block information necessary to execute a transaction / call.
+// FIXME: consider combining with `ResolvedBlockInfo`?
+#[derive(Debug, Clone, Copy)]
+pub struct BlockInfo {
+    resolved_block_number: L2BlockNumber,
+    l1_batch_timestamp_s: Option<u64>,
+}
+
+impl BlockInfo {
+    /// Fetches information for a pending block.
+    pub async fn pending(connection: &mut Connection<'_, Core>) -> anyhow::Result<Self> {
+        let resolved_block_number = connection
+            .blocks_web3_dal()
+            .resolve_block_id(api::BlockId::Number(api::BlockNumber::Pending))
+            .await
+            .map_err(DalError::generalize)?
+            .context("pending block should always be present in Postgres")?;
+        Ok(Self {
+            resolved_block_number,
+            l1_batch_timestamp_s: None,
+        })
+    }
+
+    /// Fetches information for an existing block.
+    pub async fn for_existing_block(
+        connection: &mut Connection<'_, Core>,
+        number: L2BlockNumber,
+    ) -> anyhow::Result<Self> {
+        let l1_batch = connection
+            .storage_web3_dal()
+            .resolve_l1_batch_number_of_l2_block(number)
+            .await
+            .with_context(|| format!("failed resolving L1 batch number of L2 block #{number}"))?;
+        let l1_batch_timestamp = connection
+            .blocks_web3_dal()
+            .get_expected_l1_batch_timestamp(&l1_batch)
+            .await
+            .map_err(DalError::generalize)?
+            .context("missing timestamp for non-pending block")?;
+        Ok(Self {
+            resolved_block_number: number,
+            l1_batch_timestamp_s: Some(l1_batch_timestamp),
+        })
+    }
+
+    /// Returns L2 block number.
+    pub fn block_number(&self) -> L2BlockNumber {
+        self.resolved_block_number
+    }
+
+    /// Returns L1 batch timestamp in seconds, or `None` for a pending block.
+    pub fn l1_batch_timestamp(&self) -> Option<u64> {
+        self.l1_batch_timestamp_s
+    }
+
+    fn is_pending_l2_block(&self) -> bool {
+        self.l1_batch_timestamp_s.is_none()
+    }
+
+    /// Loads historical fee input for this block. If the block is not present in the DB, returns an error.
+    pub async fn historical_fee_input(
+        &self,
+        connection: &mut Connection<'_, Core>,
+    ) -> anyhow::Result<BatchFeeInput> {
+        let header = connection
+            .blocks_dal()
+            .get_l2_block_header(self.resolved_block_number)
+            .await?
+            .context("resolved L2 block is not in storage")?;
+        Ok(header.batch_fee_input)
+    }
+
+    /// Resolves this information into [`ResolvedBlockInfo`].
+    pub async fn resolve(
+        &self,
+        connection: &mut Connection<'_, Core>,
+    ) -> anyhow::Result<ResolvedBlockInfo> {
+        let (state_l2_block_number, vm_l1_batch_number, l1_batch_timestamp);
+
+        let l2_block_header = if let Some(l1_batch_timestamp_s) = self.l1_batch_timestamp_s {
+            vm_l1_batch_number = connection
+                .storage_web3_dal()
+                .resolve_l1_batch_number_of_l2_block(self.resolved_block_number)
+                .await
+                .context("failed resolving L1 batch for L2 block")?
+                .expected_l1_batch();
+            l1_batch_timestamp = l1_batch_timestamp_s;
+            state_l2_block_number = self.resolved_block_number;
+
+            connection
+                .blocks_dal()
+                .get_l2_block_header(self.resolved_block_number)
+                .await?
+                .context("resolved L2 block disappeared from storage")?
+        } else {
+            vm_l1_batch_number = connection
+                .blocks_dal()
+                .get_sealed_l1_batch_number()
+                .await?
+                .context("no L1 batches in storage")?;
+            let sealed_l2_block_header = connection
+                .blocks_dal()
+                .get_last_sealed_l2_block_header()
+                .await?
+                .context("no L2 blocks in storage")?;
+
+            state_l2_block_number = sealed_l2_block_header.number;
+            // Timestamp of the next L1 batch must be greater than the timestamp of the last L2 block.
+            l1_batch_timestamp = seconds_since_epoch().max(sealed_l2_block_header.timestamp + 1);
+            sealed_l2_block_header
+        };
+
+        // Blocks without version specified are considered to be of `Version9`.
+        // TODO: remove `unwrap_or` when protocol version ID will be assigned for each block.
+        let protocol_version = l2_block_header
+            .protocol_version
+            .unwrap_or(ProtocolVersionId::last_potentially_undefined());
+
+        Ok(ResolvedBlockInfo {
+            state_l2_block_number,
+            state_l2_block_hash: l2_block_header.hash,
+            vm_l1_batch_number,
+            l1_batch_timestamp,
+            protocol_version,
+            is_pending: self.is_pending_l2_block(),
+        })
+    }
+}
+
+/// Resolved [`BlockInfo`] containing additional data from VM state.
+#[derive(Debug)]
+pub struct ResolvedBlockInfo {
+    state_l2_block_number: L2BlockNumber,
+    state_l2_block_hash: H256,
+    vm_l1_batch_number: L1BatchNumber,
+    l1_batch_timestamp: u64,
+    protocol_version: ProtocolVersionId,
+    is_pending: bool,
+}
+
+impl ResolvedBlockInfo {
+    /// L2 block number (as stored in Postgres). This number may differ from `block.number` provided to the VM.
+    pub fn state_l2_block_number(&self) -> L2BlockNumber {
+        self.state_l2_block_number
+    }
+}
+
+/// Arguments for VM execution necessary to set up storage and environment.
+// FIXME: rename
+#[derive(Debug, Clone)]
+pub struct TxSetupArgs {
+    // FIXME: revise fields vs args for `into_env`
+    pub execution_mode: TxExecutionMode, // changes each call
+    pub operator_account: AccountTreeId, // rarely
+    pub fee_input: BatchFeeInput,        // each call
+    pub base_system_contracts: MultiVMBaseSystemContracts, // never
+    pub validation_computational_gas_limit: u32, // rarely
+    pub chain_id: L2ChainId,             // never
+    pub enforced_base_fee: Option<u64>,  // ???
+}
+
+impl TxSetupArgs {
+    #[doc(hidden)] // Useful only for tests
+    pub fn mock(
+        execution_mode: TxExecutionMode,
+        base_system_contracts: MultiVMBaseSystemContracts,
+    ) -> Self {
+        Self {
+            execution_mode,
+            operator_account: AccountTreeId::default(),
+            fee_input: BatchFeeInput::l1_pegged(55, 555),
+            base_system_contracts,
+            validation_computational_gas_limit: u32::MAX,
+            chain_id: L2ChainId::default(),
+            enforced_base_fee: None,
+        }
+    }
+
+    pub async fn into_env(
+        self,
+        connection: &mut Connection<'_, Core>,
+        resolved_block_info: &ResolvedBlockInfo,
+    ) -> anyhow::Result<OneshotEnv> {
+        let (next_block, current_block) = load_l2_block_info(
+            connection,
+            resolved_block_info.is_pending,
+            resolved_block_info,
+        )
+        .await?;
+
+        let (system, l1_batch) = self.prepare_env(resolved_block_info, next_block);
+        let env = OneshotEnv {
+            system,
+            l1_batch,
+            current_block,
+        };
+        Ok(env)
+    }
+
+    fn prepare_env(
+        self,
+        resolved_block_info: &ResolvedBlockInfo,
+        next_block: L2BlockEnv,
+    ) -> (SystemEnv, L1BatchEnv) {
+        let Self {
+            execution_mode,
+            operator_account,
+            fee_input,
+            base_system_contracts,
+            validation_computational_gas_limit,
+            chain_id,
+            enforced_base_fee,
+        } = self;
+
+        let system_env = SystemEnv {
+            zk_porter_available: ZKPORTER_IS_AVAILABLE,
+            version: resolved_block_info.protocol_version,
+            base_system_smart_contracts: base_system_contracts
+                .get_by_protocol_version(resolved_block_info.protocol_version),
+            bootloader_gas_limit: BATCH_COMPUTATIONAL_GAS_LIMIT,
+            execution_mode,
+            default_validation_computational_gas_limit: validation_computational_gas_limit,
+            chain_id,
+        };
+        let l1_batch_env = L1BatchEnv {
+            previous_batch_hash: None,
+            number: resolved_block_info.vm_l1_batch_number,
+            timestamp: resolved_block_info.l1_batch_timestamp,
+            fee_input,
+            fee_account: *operator_account.address(),
+            enforced_base_fee,
+            first_l2_block: next_block,
+        };
+        (system_env, l1_batch_env)
+    }
+}
+
+async fn load_l2_block_info(
+    connection: &mut Connection<'_, Core>,
+    is_pending_block: bool,
+    resolved_block_info: &ResolvedBlockInfo,
+) -> anyhow::Result<(L2BlockEnv, Option<StoredL2BlockEnv>)> {
+    let mut current_block = None;
+    let next_block = read_stored_l2_block(connection, resolved_block_info.state_l2_block_number)
+        .await
+        .context("failed reading L2 block info")?;
+
+    let next_block = if is_pending_block {
+        L2BlockEnv {
+            number: next_block.number + 1,
+            timestamp: resolved_block_info.l1_batch_timestamp,
+            prev_block_hash: resolved_block_info.state_l2_block_hash,
+            // For simplicity, we assume each L2 block create one virtual block.
+            // This may be wrong only during transition period.
+            max_virtual_blocks_to_create: 1,
+        }
+    } else if next_block.number == 0 {
+        // Special case:
+        // - For environments, where genesis block was created before virtual block upgrade it doesn't matter what we put here.
+        // - Otherwise, we need to put actual values here. We cannot create next L2 block with block_number=0 and `max_virtual_blocks_to_create=0`
+        //   because of SystemContext requirements. But, due to intrinsics of SystemContext, block.number still will be resolved to 0.
+        L2BlockEnv {
+            number: 1,
+            timestamp: 0,
+            prev_block_hash: L2BlockHasher::legacy_hash(L2BlockNumber(0)),
+            max_virtual_blocks_to_create: 1,
+        }
+    } else {
+        // We need to reset L2 block info in storage to process transaction in the current block context.
+        // Actual resetting will be done after `storage_view` is created.
+        let prev_block_number = resolved_block_info.state_l2_block_number - 1;
+        let prev_l2_block = read_stored_l2_block(connection, prev_block_number)
+            .await
+            .context("failed reading previous L2 block info")?;
+
+        let mut prev_block_hash = connection
+            .blocks_web3_dal()
+            .get_l2_block_hash(prev_block_number)
+            .await
+            .map_err(DalError::generalize)?;
+        if prev_block_hash.is_none() {
+            // We might need to load the previous block hash from the snapshot recovery metadata
+            let snapshot_recovery = connection
+                .snapshot_recovery_dal()
+                .get_applied_snapshot_status()
+                .await
+                .map_err(DalError::generalize)?;
+            prev_block_hash = snapshot_recovery.and_then(|recovery| {
+                (recovery.l2_block_number == prev_block_number).then_some(recovery.l2_block_hash)
+            });
+        }
+
+        current_block = Some(prev_l2_block);
+        L2BlockEnv {
+            number: next_block.number,
+            timestamp: next_block.timestamp,
+            prev_block_hash: prev_block_hash.with_context(|| {
+                format!("missing hash for previous L2 block #{prev_block_number}")
+            })?,
+            max_virtual_blocks_to_create: 1,
+        }
+    };
+
+    Ok((next_block, current_block))
+}
+
+async fn read_stored_l2_block(
+    connection: &mut Connection<'_, Core>,
+    l2_block_number: L2BlockNumber,
+) -> anyhow::Result<StoredL2BlockEnv> {
+    let l2_block_info_key = StorageKey::new(
+        AccountTreeId::new(SYSTEM_CONTEXT_ADDRESS),
+        SYSTEM_CONTEXT_CURRENT_L2_BLOCK_INFO_POSITION,
+    );
+    let l2_block_info = connection
+        .storage_web3_dal()
+        .get_historical_value_unchecked(l2_block_info_key.hashed_key(), l2_block_number)
+        .await?;
+    let (l2_block_number_from_state, timestamp) = unpack_block_info(h256_to_u256(l2_block_info));
+
+    let l2_block_txs_rolling_hash_key = StorageKey::new(
+        AccountTreeId::new(SYSTEM_CONTEXT_ADDRESS),
+        SYSTEM_CONTEXT_CURRENT_TX_ROLLING_HASH_POSITION,
+    );
+    let txs_rolling_hash = connection
+        .storage_web3_dal()
+        .get_historical_value_unchecked(l2_block_txs_rolling_hash_key.hashed_key(), l2_block_number)
+        .await?;
+
+    Ok(StoredL2BlockEnv {
+        number: l2_block_number_from_state as u32,
+        timestamp,
+        txs_rolling_hash,
+    })
+}
