+++ core/node/api_server/src/execution_sandbox/validate.rs
@@ -11,13 +11,14 @@ use zksync_types::{
     api::state_override::StateOverride, l2::L2Tx, Address, TRUSTED_ADDRESS_SLOTS,
     TRUSTED_TOKEN_SLOTS,
 };
+use zksync_vm_executor::oneshot::TxSetupArgs;
 
 use super::{
     apply,
     execute::TransactionExecutor,
     storage::StorageWithOverrides,
     vm_metrics::{SandboxStage, EXECUTION_METRICS, SANDBOX_METRICS},
-    BlockArgs, TxSetupArgs, VmPermit,
+    BlockArgs, VmPermit,
 };
 
 /// Validation error used by the sandbox. Besides validation errors returned by VM, it also includes an internal error
@@ -39,20 +40,25 @@ impl TransactionExecutor {
         tx: L2Tx,
         setup_args: TxSetupArgs,
         block_args: BlockArgs,
-        computational_gas_limit: u32,
+        whitelisted_tokens_for_aa: &[Address],
     ) -> Result<(), ValidationError> {
         let total_latency = SANDBOX_METRICS.sandbox[&SandboxStage::ValidateInSandbox].start();
         let validation_params = get_validation_params(
             &mut connection,
             &tx,
-            computational_gas_limit,
-            &setup_args.whitelisted_tokens_for_aa,
+            setup_args.validation_computational_gas_limit,
+            whitelisted_tokens_for_aa,
         )
         .await
         .context("failed getting validation params")?;
 
-        let (env, storage) =
-            apply::prepare_env_and_storage(connection, setup_args, &block_args).await?;
+        let (env, storage) = apply::prepare_env_and_storage(
+            connection,
+            setup_args,
+            &block_args,
+            self.storage_caches(),
+        )
+        .await?;
         let storage = StorageWithOverrides::new(storage, &StateOverride::default());
 
         let stage_latency = SANDBOX_METRICS.sandbox[&SandboxStage::Validation].start();
